detect-requirements.txt
Pillow
imgaug
numpy
pandas
pyyaml
torch==1.13.1
tqdm
